more single linked lists

+++ 0x13-more_singly_linked_lists/101-print_listint_safe.c
@@ -9,7 +9,7 @@
 size_t print_listint_safe(const listint_t *head)
 {
 	size_t num = 0;
-	unsigned int d;
+	long int d;
 
 	while (head)
 	{
